feat(radar): make the MR board a tier; radar becomes a reconcile engine
#527 introduced ops.radar.radar_tiers and stopped half way — the GitLab
MR board stayed hardcoded in main(). Radar's core is "reconcile
registered tiers against live truth, heal their watchers, report, stay
idempotent, and never render an unknown as green", and merge requests
appear nowhere in it. GitLab MRs were the first tier anyone wrote.

The board moves to presets/watch/tiers/gl_mrs.py unchanged in behaviour.
gl-runners loses the watcher radar was hardcoding on its behalf and
declares its own. With no tiers configured radar refuses on stderr with
exit 1 and names the fix, rather than printing nothing — an unconfigured
radar that is silent is byte-identical to a healthy one.

BREAKING CHANGE: add {"radar_tiers": {"gl-mrs": {}}} to the radar op
block to keep the board you had.

The contract:
  radar_report(options) -> (lines, healthy)   as before
  RADAR_QUIET_DEFAULT                          new; False for a tier
                                               whose report is the board
  options["_arg"]                              raw invocation argument
  options["_watch"]                            radar's bounded spawner

radar_watchers() as proposed was not implemented: a declared slot must
be spawned before the report, and the MR tier must not spawn its feed
when GitLab was unreachable. Only the tier knows when spawning is safe,
and it needs the result inside its own report. Radar owns the bound, the
tier owns the timing.

A tier that raises or cannot be resolved goes to stderr with exit 1, on
a channel separate from the board — one broken tier cannot cost another
its board, and cannot leave radar exiting 0. Three states, not two.

Closes #528
