+++ presets/watch.json
@@ -22,7 +22,7 @@
     "radar": {
       "cmd": "{python} {path}watch/radar.py {args}",
       "timeout": 90,
-      "description": "Reconcile watch coverage against live GitLab, then report. Live MRs are authoritative; state files are cache. Prunes terminal state files, flags pipeline drift (last_event vs source_state), respawns watchers for open MRs with no live poller, and ensures one live gitlab-mr-feed poller so MRs opened after this run are still discovered. Full board on cold start, delta-only after. Idempotent — safe on every session start. Takes the gl-mrs filter vocabulary; the board, the watcher fleet and the feed all come from that one filter, and each filter keeps its own snapshot. A repeated key (author=@me,author=modular.system) fans out into one query per value, unioned by iid. Standing exclusions live in ops.radar.radar_exclusions ({iid: reason} or {iid: {reason, until}}): they keep a permanently-red row off the board without hiding it — the footer counts them, one line names each with its current status and reason, and the suppression lifts by itself once the MR is no longer red. Exclusions never narrow the watcher fleet, the feed or the event filter. The filter is per-invocation and is not persisted, so the footer names the scope the board was built from on every run, default included, and a feed poller still live on another scope is named too — its MRs are not on this board.",
+      "description": "Reconcile registered radar tiers against live truth, then report. Tiers are registered in ops.radar.radar_tiers (reaches the preset as SUPERTOOL_RADAR_TIERS) and there is no default: with none configured radar refuses and names the fix, because an unconfigured radar that prints nothing is byte-identical to a healthy one. `{\"gl-mrs\": {}}` is the GitLab MR board — live MRs are authoritative, state files are cache; it prunes terminal state files, flags pipeline drift (last_event vs source_state), respawns watchers for open MRs with no live poller, and keeps one gitlab-mr-feed poller alive so MRs opened after this run are still discovered. `{\"gl-runners\": {}}` adds the runner fleet. Full board on cold start, delta-only after. Idempotent — safe on every session start. Takes the gl-mrs filter vocabulary; the board, the watcher fleet and the feed are always one resolved filter. Standing exclusions live in ops.radar.radar_exclusions.",
       "syntax": "radar[:author=@me,author=modular.system,state=opened]"
     }
   }
