+++ presets/watch/radar.py
@@ -1,134 +1,89 @@
 #!/usr/bin/env python3
-"""radar — reconcile watch coverage against live GitLab, then report.
+"""radar — reconcile registered tiers against live truth, then report.
 
-`watches` reports that pollers are alive. It cannot report what is true, and
-the two diverge routinely:
+Radar's core is one sentence, and merge requests are not in it:
 
-    last_event    : pipeline_failed  on pipeline 154177
-    source_state  : running          on pipeline 154180
+    Reconcile registered tiers against live truth, heal their watchers, report,
+    stay idempotent, and never render an unknown as green.
 
-`source_state` is the truth; `last_event` is history. A board built on
-`last_event` calls that MR broken when it is mid-retry on a newer pipeline.
+GitLab MRs were simply the first tier anyone wrote (#528). The board lives in
+`tiers/gl_mrs.py` and joins on exactly the same footing as `gl-runners`,
+`gh-prs`, or anything else with a population and a standing question.
 
-Events also cannot survive a session boundary. The transport is
-fire-and-forget, so an event emitted with no listener is gone permanently;
-pollers are processes that die with the machine; and pollers stop themselves
-on terminal state by design. At the start of a session an event-driven view
-therefore knows nothing — and "knows nothing" renders identically to
-"everything is green". That is the failure this op exists to remove, so state
-is the floor and events are the optimisation, not the other way round.
+    {"ops": {"radar": {"radar_tiers": {"gl-mrs": {}, "gl-runners": {}}}}}
 
-Hence a reconcile, not a printer:
+With no tiers configured radar refuses and says how to fix it. It does not
+quietly do nothing, and it does not assume `gl-mrs`:
 
-  1. live truth   one gl-mrs query for open MRs — authoritative. The state
-                  files are cache and may be absent or hours stale.
-  2. reconcile    prune state files whose watcher reached a terminal state;
-                  flag drift where the last event fired on an older pipeline
-                  than the one now running.
-  3. heal         respawn a watcher for every open MR with no live poller.
-                  Covers reboot, crash, a cleared /tmp, and MRs that were
-                  green when watchers were last spawned — in one step.
-  3b. feed        ensure one live gitlab-mr-feed poller. Reconcile is a
-                  snapshot; the feed is the thing that keeps discovering after
-                  radar returns, so a session idle for an hour still learns
-                  about an MR opened fifty minutes ago. Its absence is
-                  reported loudly, because a feed that is not running looks
-                  exactly like a day on which nothing happened.
-  4. report       full board on cold start, delta-only afterwards.
+  * Silence is the failure this whole preset is built against. An unconfigured
+    radar that prints nothing is byte-identical to a healthy one.
+  * A `gl-mrs` default is an opinion imposed on strangers — it points GitLab
+    API calls at people who may be on GitHub, and hides from them that radar
+    is configurable at all.
+  * The refusal is also the documentation: it teaches the config at the moment
+    someone needs it.
 
-The population is an argument, in the same filter vocabulary as `gl-mrs`:
+The tier contract
+-----------------
 
-    radar                            defaults.DEFAULT_FILTER
-    radar:author=modular.system      what the automated agent opened
-    radar:author=@me,author=x        two queries, unioned by iid
+A tier is a Python module reachable by name (see `_tier_module`) exposing:
 
-One filter, one population, one board. `live_open_mrs`, `heal` and the feed
-poller are three views of a single resolved filter and are never derived from
-different ones — a board that omits MRs it is actively receiving events for
-renders exactly like a board where those MRs are fine, which is the same
-silent incompleteness this op exists to remove.
+    RADAR_OPTIONS       set[str] — config keys this tier understands. Anything
+                        else in its config block is named on the board, never
+                        silently ignored.
+    RADAR_QUIET_DEFAULT bool, optional (default True) — is a healthy tier
+                        silent? True for a side-concern like the runner fleet,
+                        where a green line per run is the noise that trains a
+                        reader to skim past the red one. False for a tier whose
+                        report *is* the board: an MR reconcile that prints
+                        nothing on a quiet day is indistinguishable from one
+                        that failed to run.
+    radar_report(options) -> (lines, healthy)
+                        `healthy` means "this tier could tell you the truth",
+                        not "the world is fine". A board full of red MRs is
+                        healthy; a board that could not be built is not.
 
-That filter lives for one invocation and is not persisted, so a session that
-widened the board and then runs a bare `radar` silently gets the default
-population back (#486). The footer therefore names the scope on every board,
-the default one included: "no label" used to spell both "this is the default"
-and "nobody said". A feed poller still live on another scope is named too —
-changing the filter respawns the feed without retiring the old one, so
-effective scope splits between what this call passed and what a still-running
-watcher was started with. It is reported, not killed: two populations at once
-is legitimate (see `prune_terminal`), and only the reader can say which one
-they meant. What this deliberately does not give is continuity — re-widening
-still means re-typing the filter, because a population read from a file
-nobody in the session chose is the hidden state that caused the split.
+Radar injects two reserved keys into `options` before the call. Config cannot
+set them — `read_tiers` refuses any key starting with `_` — so a tier can trust
+them:
 
-The snapshot is keyed by that filter for the same reason. Two populations
-sharing one snapshot file would report every MR of the first as new and every
-MR of the second as gone; a lying delta is worse than no delta.
+    _arg    str, the raw invocation argument (`radar:author=@me` -> "author=@me")
+    _watch  callable(source, scope, only=None) -> "alive"|"spawned"|"failed"|
+            "capped". Radar's bounded spawner. Every slot a tier asks for is
+            recorded, and radar itself emits the cap warning when one is
+            refused.
 
-Idempotent, so it is safe on every session start and on a loop. That includes
-the feed: a live PID short-circuits the spawn, so N radar runs still leave one
-feed poller and one copy of every discovery event.
+`_watch` is a callable rather than the declarative `radar_watchers(options) ->
+[(source, scope, only)]` #528 proposed, and the difference is load-bearing. A
+declared slot must be spawned before the report, and the MR tier must *not*
+spawn its discovery feed when live GitLab was unreachable — the current
+behaviour, pinned by a test. Only the tier knows whether spawning is safe, and
+it needs the spawn result inside its own report (the feed's status is a token
+in the board footer). Two mechanisms for one job is the drift this codebase
+keeps filing bugs about, so there is one: radar owns the bound, the tier owns
+the timing.
 
-"Nothing moved" means: the set of open MRs is unchanged, no MR changed
-pipeline status / pipeline id / draft / conflict flag, and radar itself took
-no action (nothing pruned, nothing healed, no drift). Standing failures and
-conflicts are re-printed even when unchanged — an unfixed red is a current
-fact, not history. When nothing moved radar still prints one summary line
-rather than nothing at all: total silence is indistinguishable from a radar
-that failed to run.
+Never green when it cannot tell
+-------------------------------
 
-Standing exclusions
--------------------
-
-Some MRs are red for a reason nobody intends to fix soon, and re-printing
-that row every run is how a reader learns to skim a board — which is how a
-real red gets missed. `ops.radar.radar_exclusions` in `.supertool.json` moves
-that suppression out of the reader's memory and into the tool.
-
-An exclusion is the one thing in this op that can hide a failure, so it is
-built to be the *opposite* of a silent omission at four points:
-
-  accounted   the row goes, the MR does not. The footer carries an
-              `N excluded` token and one line names each suppressed MR, its
-              current status and the configured reason. Tallies describe the
-              board that was printed, so `1 failing` is never a count with no
-              row behind it.
-  reasoned    an exclusion with no reason is refused and the row renders. The
-              field that makes it auditable is the field that makes it work.
-  self-expiring
-              an exclusion only ever suppresses a *standing problem*. The
-              moment the MR goes green and unconflicted the suppression lifts
-              by itself and the board says the reason is spent — so an
-              exclusion cannot outlive what it was written for. An optional
-              `until` date expires it on a schedule as well.
-  board-only  an exclusion is a statement about one row, not a change of
-              population. The watcher fleet, the feed and the event filter are
-              untouched, deliberately breaking the one-filter symmetry: the
-              filter says what radar is responsible for, and radar does not
-              stop being responsible for an MR because its row is noisy. The
-              watcher is also the only thing that can still report a push
-              while the row is suppressed, which is exactly when an exclusion
-              is most likely to have gone stale.
-
-Every unanswerable case — unparseable config, unknown key shape, an iid that
-is not in the population — resolves to *show the row*.
+Three states, not two: ok, a finding, and *cannot tell*. A tier that fails to
+resolve or raises is not a quiet degradation — its message goes to stderr and
+radar exits non-zero, because a refactor that traded a loud failure for a quiet
+one would be the house defect (#486, #533) delivered by the fix for it.
 """
 from __future__ import annotations
 
-import datetime
-import glob
-import hashlib
 import importlib.util
 import json
 import os
 import sys
+from collections.abc import Callable
 from pathlib import Path
-from typing import Any
 
 _HERE = Path(__file__).parent
 
 sys.path.insert(0, str(_HERE))
-import defaults  # noqa: E402
+import dispatcher  # noqa: E402
 import transport  # noqa: E402
 
 
@@ -140,406 +95,38 @@ def _load(name: str, path: Path):
     return module
 
 
-mrs = _load("radar_gitlab_mrs", _HERE.parents[1] / "presets" / "gitlab" / "mrs.py")
-runners_op = _load("radar_gitlab_runners", _HERE.parents[1] / "presets" / "gitlab" / "runners.py")
-dispatcher = _load("radar_watch_dispatcher", _HERE / "dispatcher.py")
-
-SOURCE = defaults.DEFAULT_SOURCE
-FEED_SOURCE = defaults.DEFAULT_FEED_SOURCE
-FEED_SCOPE = defaults.DEFAULT_FEED_SCOPE
-SNAPSHOT_PREFIX = "supertool-radar"
-
-# The fleet tier. Scope-free, unlike the feed: there is one runner fleet behind
-# every filter, so this is not keyed by the board's population.
-FLEET_SOURCE = "gl-runners"
-FLEET_SCOPE = "fleet"
-
 # ops.radar.radar_tiers, JSON-encoded into the env by the op runner — the same
-# route ops.radar.radar_exclusions takes.
+# route ops.gl-job.job_patterns takes.
 TIERS_ENV = "SUPERTOOL_RADAR_TIERS"
 
-# A registered tier that also wants a background watcher names it here. Only
-# consulted for tiers actually registered, so an unregistered tier costs no
-# poller and no API call.
-_TIER_WATCH_SOURCES = {"gl-runners": (FLEET_SOURCE, FLEET_SCOPE)}
+# Keys radar fills in on a tier's options dict. Refused in config so a tier can
+# tell radar's context apart from the operator's configuration.
+RESERVED_PREFIX = "_"
 
-# ops.radar.radar_exclusions in .supertool.json, JSON-encoded into the env by
-# the op runner — the same route ops.gl-job.job_patterns takes.
-EXCLUSIONS_ENV = "SUPERTOOL_RADAR_EXCLUSIONS"
+NO_TIERS = ('radar: no tiers configured. Add ops.radar.radar_tiers to .supertool.json —\n'
+            '       e.g. {"gl-mrs": {}} for the GitLab MR board.')
 
-_FIX_HINT = "remove it from ops.radar.radar_exclusions in .supertool.json"
-
-FEED_LABEL = {"alive": "feed ok", "spawned": "feed respawned", "failed": "feed DOWN",
-              "capped": "feed DOWN (respawn capped)"}
-
-
-class RadarError(RuntimeError):
-    """Live GitLab could not be reached. Never degrade to 'all green'."""
-
-
-def default_filter() -> dict[str, list[str]]:
-    """The population a bare `radar` covers, read from defaults.py.
-
-    Not a literal here: the whole point of that module is that the shell
-    supervisor and this op cannot describe different populations.
-    """
-    return mrs._parse_multi(defaults.DEFAULT_FILTER)[0]
-
-
-def resolve_filter(argv: list[str] | None = None) -> dict[str, list[str]]:
-    """The one filter every other step reads. gl-mrs vocabulary, or default.
-
-    A key may repeat — `author=@me,author=modular.system` — because GitLab
-    takes one author per query and the union is the population the caller
-    described.
-    """
-    arg = argv[1].strip() if argv and len(argv) > 1 and argv[1] else ""
-    multi = mrs._parse_multi(arg)[0] if arg else {}
-    return multi or default_filter()
-
-
-def filter_string(multi: dict[str, list[str]]) -> str:
-    """The filter back in gl-mrs arg form — what the user typed, normalised."""
-    return ",".join(f"{k}={v}" for k, vals in multi.items() for v in vals)
-
-
-def canonical_filter_string(multi: dict[str, list[str]]) -> str:
-    """The filter in one fixed spelling — sorted keys, sorted values, deduped.
-
-    Used where the filter string is an *identity*, not a display: the feed
-    watcher id is the pid filename, so `author=a,author=b` and
-    `author=b,author=a` would otherwise be two pollers over one population,
-    i.e. two copies of every mr_opened (#476).
-
-    Safe as an identity key in the direction that matters. It only merges
-    filters that are already the same set, so it can never refuse to start a
-    poller for a filter that would have selected something different — the
-    failure that would show up as a watcher silently not existing.
-    """
-    return ",".join(f"{k}={v}"
-                    for k, vals in sorted(multi.items())
-                    for v in sorted(set(vals)))
-
-
-def filter_key(multi: dict[str, list[str]]) -> str:
-    """Stable short hash of the filter, insensitive to key and value order.
-
-    `author=a,author=b` and `author=b,author=a` are the same population and
-    must share a snapshot; `author=a` and `author=b` must not.
-    """
-    norm = {k: sorted(set(v)) for k, v in sorted(multi.items())}
-    blob = json.dumps(norm, sort_keys=True).encode("utf-8")
-    return hashlib.sha1(blob).hexdigest()[:12]
-
-
-def _snapshot_path(multi: dict[str, list[str]]) -> str:
-    return os.path.join(transport.STATE_DIR,
-                        f"{SNAPSHOT_PREFIX}.{filter_key(multi)}.snapshot.json")
-
-
-# ---------------------------------------------------------------------------
-# 1. live truth
-# ---------------------------------------------------------------------------
-
-def _query(filters: dict[str, str], per_page: int) -> list[dict]:
-    """One `glab mr list`. RadarError on any failure, never an empty list.
-
-    A partial union is a board that is quietly missing rows, so one failing
-    query fails the whole reconcile — an unreachable GitLab must never render
-    as an MR that is fine.
-    """
-    try:
-        result = mrs._run(mrs._build_list_cmd(filters, per_page))
-    except Exception as exc:  # noqa: BLE001 — surfaced as RadarError
-        raise RadarError(f"glab mr list failed: {exc}") from exc
-    if result.returncode != 0:
-        err = (result.stderr or "").strip() or "unknown error"
-        if "not logged in" in err.lower() or "401" in err:
-            raise RadarError("glab not authenticated. Run: glab auth login")
-        raise RadarError(f"glab mr list: {err}")
-    try:
-        data = json.loads(result.stdout)
-    except json.JSONDecodeError as exc:
-        raise RadarError("could not parse glab JSON output") from exc
-    if not isinstance(data, list):
-        raise RadarError("glab returned no MR list")
-    return data
-
-
-def live_open_mrs(multi: dict[str, list[str]] | None = None) -> list[dict]:
-    """Every open MR the filter describes, pipeline-enriched.
-
-    One gl-mrs query per value combination, unioned by iid — the list endpoint
-    takes a single author, so two authors is two calls. Enrichment runs once
-    over the union, so an MR both queries return is enriched once.
-    """
-    multi = default_filter() if multi is None else multi
-    cfg = mrs._get_config()
-    merged: dict[str, dict] = {}
-    for filters in mrs._expand_filters(multi):
-        for m in _query(filters, cfg["per_page"]):
-            if isinstance(m, dict) and m.get("iid") is not None:
-                merged.setdefault(str(m["iid"]), m)
-    data = list(merged.values())
-    mrs._enrich(data, cfg["enrich_cap"], cfg["enrich_workers"])
-    return data
-
-
-# ---------------------------------------------------------------------------
-# 2. reconcile
-# ---------------------------------------------------------------------------
-
-def read_state_files() -> dict[str, dict]:
-    """{iid: full state file} for this source, from the state-file cache."""
-    prefix = f"supertool-watch-{SOURCE}__"
-    suffix = ".state.json"
-    out: dict[str, dict] = {}
-    pattern = os.path.join(transport.STATE_DIR, f"{prefix}*{suffix}")
-    for path in sorted(glob.glob(pattern)):
-        iid = os.path.basename(path)[len(prefix):-len(suffix)]
-        if not iid:
-            continue
-        try:
-            with open(path, encoding="utf-8") as f:
-                loaded = json.load(f)
-        except (OSError, json.JSONDecodeError):
-            continue
-        if isinstance(loaded, dict):
-            out[iid] = loaded
-    return out
-
-
-def prune_terminal(states: dict[str, dict], watched: set[str]) -> list[str]:
-    """Delete state files whose watcher reached a terminal state (item 4).
-
-    A file owned by a live poller is left alone — the poller will clear it
-    itself when it stops. Only MRs whose live state is terminal are pruned;
-    "absent from my open list" is deliberately not a prune trigger, because a
-    watcher may legitimately follow an MR this radar's filter never returns —
-    a second radar over another population is exactly that case.
-    """
-    poller = dispatcher._load_source(SOURCE)
-    is_terminal = getattr(poller, "is_terminal", None) if poller else None
-    if is_terminal is None:
-        return []
-    pruned: list[str] = []
-    for iid, full in states.items():
-        if iid in watched:
-            continue
-        if is_terminal(full.get("source_state") or {}):
-            if transport.clear_state(SOURCE, iid):
-                pruned.append(iid)
-    return pruned
-
-
-def drift(states: dict[str, dict]) -> dict[str, tuple[str, str]]:
-    """{iid: (event_pipeline_id, current_pipeline_id)} where they disagree.
-
-    A newer pipeline superseded the last event, so the event is stale history.
-    Reported rather than silently resolved — the divergence is the signal.
-    """
-    out: dict[str, tuple[str, str]] = {}
-    for iid, full in states.items():
-        event_pipe = str(((full.get("last_event") or {}).get("payload") or {}).get("pipeline_id") or "")
-        live_pipe = str((full.get("source_state") or {}).get("pipeline_id") or "")
-        if event_pipe and live_pipe and event_pipe != live_pipe:
-            out[iid] = (event_pipe, live_pipe)
-    return out
-
-
-# ---------------------------------------------------------------------------
-# 3. heal
-# ---------------------------------------------------------------------------
-
-def heal(open_iids: list[str], watched: set[str]) -> tuple[list[str], list[str], list[str]]:
-    """Respawn a watcher for every open MR without a live poller.
-
-    Returns (healed, still_uncovered, refused); `refused` is a subset of
-    `still_uncovered`. The event filter comes from defaults.py
-    so a healed watcher is identical to one watch-mine.sh would have spawned.
-
-    Spawning goes through `dispatcher.start_poller`, the one door, for the
-    reason #476 gives: `watched` is derived from pidfiles a grandchild
-    publishes after a fork, an import and a detach, so a caller that reads it
-    and then forks is testing a slot it cannot see being filled. Radar runs on
-    a loop from more than one place, and item 1 of #417 widened this tier from
-    "the MRs that are already red" to "every open MR" — so this is the tier
-    where losing that race duplicates the whole fleet rather than one poller.
-
-    A slot already held is neither healed nor uncovered: radar did not spawn
-    it, so claiming the action would be false, but the MR *is* covered, and a
-    spurious "unwatched" warning corrodes the only signal on the board that
-    has to be trusted absolutely.
-
-    Healing is bounded, and the bound is #513's substance. A gap left by a
-    poller that died is reaped first, so the death is on record before the
-    claim overwrites the evidence; past `transport.DEATH_RESPAWN_LIMIT` deaths
-    the slot is refused rather than respawned. Respawning forever would keep
-    the board green while a watcher failed over and over — a visible failure
-    converted into an invisible loop, which is the same bug one level up. A
-    refused slot is reported as uncovered, because it is: the automation has
-    stopped, and the only thing worse than saying so is not saying so.
-    """
-    gaps = [iid for iid in open_iids if iid not in watched]
-    if not gaps:
-        return [], [], []
-    if dispatcher._load_source(SOURCE) is None:
-        return [], gaps, []
-    only = [e for e in defaults.DEFAULT_ONLY.split(",") if e]
-    healed: list[str] = []
-    failed: list[str] = []
-    refused: list[str] = []
-    for iid in gaps:
-        transport.reap_dead_pidfile(SOURCE, iid)
-        if len(transport.deaths(SOURCE, iid)) >= transport.DEATH_RESPAWN_LIMIT:
-            refused.append(iid)
-            continue
-        status, _pid = dispatcher.start_poller(SOURCE, iid, only)
-        if status == "spawned":
-            healed.append(iid)
-        elif status == "failed":
-            failed.append(iid)
-    return healed, failed + refused, refused
-
-
-def loss_warnings(healed: list[str], refused: list[str]) -> list[str]:
-    """What the board says about watchers it lost, and about ones it gave up on.
-
-    Two lines, two different actions. A healed loss is reported once, on the
-    run that healed it, and then goes quiet — a permanent mark on a slot that
-    is now covered is what trains a reader to skim. A refusal is reported on
-    every run until an operator re-arms or acknowledges it, because the MR is
-    genuinely unwatched for as long as it stands.
-    """
-    out: list[str] = []
-    for iid in refused:
-        n = len(transport.deaths(SOURCE, iid))
-        out.append(f"radar: WARNING — !{iid} has lost its poller {n} times; "
-                   f"NOT respawning. This MR is unwatched until the cause is "
-                   f"fixed and it is re-armed: "
-                   f"./supertool 'watch:{SOURCE}:{iid}'.")
-    for iid in healed:
-        recorded = transport.deaths(SOURCE, iid)
-        if not recorded:
-            continue
-        last = recorded[-1].get("pid", "?")
-        out.append(f"radar: NOTE — !{iid} lost its poller (PID {last} died "
-                   f"without being unwatched, {len(recorded)} recorded); "
-                   f"respawned.")
-    return out
-
-
-# ---------------------------------------------------------------------------
-# 3b. feed — the tier that discovers MRs radar has never seen
-# ---------------------------------------------------------------------------
-
-def feed_scope(multi: dict[str, list[str]] | None = None) -> str:
-    """The feed watcher id covering this population.
-
-    The feed source accepts either one of its aliases or a literal gl-mrs
-    filter string as its id, so the board's filter reaches the discovery tier
-    unchanged. An alias is preferred when it expands to the same filter: the
-    id is the pid filename, so `@me` and `author=@me,state=opened` would
-    otherwise be two pollers over one population, i.e. two copies of every
-    mr_opened. For the same reason the fallback is the canonical spelling of
-    the filter rather than the caller's: key order is not identity (#476).
-    """
-    multi = default_filter() if multi is None else multi
-    poller = dispatcher._load_source(FEED_SOURCE)
-    aliases = getattr(poller, "ALIASES", None) or {}
-    for alias, expansion in aliases.items():
-        if mrs._parse_multi(expansion)[0] == multi:
-            return alias
-    return canonical_filter_string(multi)
-
-
-def feed_pid(scope: str = FEED_SCOPE) -> int:
-    """PID recorded for the feed poller, or 0 when there is no readable file."""
-    return transport.read_pid(FEED_SOURCE, scope)
-
-
-def ensure_feed(scope: str = FEED_SCOPE) -> str:
-    """Guarantee exactly one live feed poller. "alive" | "spawned" | "failed".
-
-    Radar is idempotent and run on a loop, so the feed must be too: a live PID
-    short-circuits before any spawn. Without that check every radar run would
-    stack another feed poller, and n pollers over one filter means n copies of
-    every mr_opened.
-
-    Reading the PID and then spawning was not that check — the poller publishes
-    its PID after a fork and a detach, and radar runs on a 5-minute loop from
-    more than one place, so two runs landing in that window both saw an empty
-    slot and both spawned. That is where the same-second groups of identical
-    `radar.py` processes in #476 came from: a feed poller is forked from radar,
-    so it wears radar's argv. `start_poller` closes the window by claiming the
-    slot atomically before the fork.
-    """
-    if dispatcher._load_source(FEED_SOURCE) is None:
-        return "failed"
-    only = [e for e in defaults.DEFAULT_FEED_ONLY.split(",") if e]
-    return ensure_watcher(FEED_SOURCE, scope, only)
-
-
-def other_feed_scopes(scope: str = FEED_SCOPE) -> list[str]:
-    """Live feed pollers covering a population other than this board's.
-
-    Changing the filter respawns the feed; the previous one is not retired, so
-    a machine can carry a feed started with `author=@me,author=x` while the
-    current invocation resolved plain `author=@me`. Effective scope is then
-    split between what this call passed and what a still-running watcher was
-    started with, and neither half is the whole — a board that reports its own
-    half as the answer is narrower than the reader believes it is.
-
-    Read-only: pid files only, nothing is spawned and nothing is killed. Two
-    scopes are a legitimate arrangement (`prune_terminal` says as much about
-    per-MR watchers), so this reports the split rather than resolving it.
-    """
-    return sorted({
-        str(row.get("id") or "")
-        for row in transport.list_active_pids()
-        if row.get("source") == FEED_SOURCE and str(row.get("id") or "") != scope
-    } - {""})
-
-
-def feed_error(scope: str = FEED_SCOPE) -> str:
-    """Last error the feed poller recorded, or "" when it is polling cleanly.
-
-    A feed that is alive but erroring every tick discovers nothing while
-    looking healthy in `watches` — the same silence as a dead one, so it gets
-    the same report. The dispatcher clears this key on a successful poll, so a
-    message here is current rather than a scar.
-    """
-    state = transport.read_state(FEED_SOURCE, scope)
-    return str((state.get("last_error") or {}).get("message") or "")
-
-
-# ---------------------------------------------------------------------------
-# 3c. fleet — the tier that says whether the board is even the right question
-# ---------------------------------------------------------------------------
 
 def ensure_watcher(source: str, scope: str, only: list[str] | None = None) -> str:
     """One live poller for a slot. "alive"|"spawned"|"failed"|"capped".
 
-    Same idempotence contract as the per-MR heal, for the same reason: radar
-    runs on a loop, and n pollers over one slot means n copies of every event.
-    `start_poller` claims the slot before the fork.
+    Idempotent, because radar runs on a loop and n pollers over one slot means
+    n copies of every event. `start_poller` claims the slot before the fork:
+    reading a pidfile and then spawning is not that check, since the PID is
+    published after a fork and a detach, so two runs landing in that window
+    both see an empty slot and both spawn (#476).
 
-    The death cap from #513 is applied here rather than only in `heal`. That
-    bound lived on the per-MR path alone, so every other spawner — the feed,
-    and any tier watcher — would respawn a poller that dies on every tick,
-    forever, silently. Reproducing a fixed defect in a new tier is how it comes
-    back, so the cap belongs at the one place that spawns.
+    The death cap from #513 is applied here rather than in any one caller. That
+    bound lived on the per-MR path alone, so every other spawner would respawn
+    a poller that dies on every tick, forever, silently. Reproducing a fixed
+    defect in a new tier is how it comes back, so the cap belongs at the one
+    place that spawns.
     """
     if dispatcher._load_source(source) is None:
         return "failed"
     if len(transport.deaths(source, scope)) >= transport.DEATH_RESPAWN_LIMIT:
         return "capped"
     return dispatcher.start_poller(source, scope, only or [])[0]
-
-
-def ensure_fleet() -> str:
-    """Back-compat alias for the fleet tier's watcher."""
-    return ensure_watcher(FLEET_SOURCE, FLEET_SCOPE)
 
 
 def watcher_cap_warnings(statuses: dict[str, str]) -> list[str]:
@@ -559,14 +146,13 @@ def watcher_cap_warnings(statuses: dict[str, str]) -> list[str]:
 def read_tiers(raw: str | None = None) -> tuple[dict[str, dict], list[str]]:
     """({op_name: options}, complaints) from ops.radar.radar_tiers.
 
-    Empty by default, and that default is the point. Radar is an MR reconcile
-    tool for most of the people who install it; a tier that reaches for runners,
-    or any other privileged resource, must be asked for. Shipping one on by
-    default would hand every stranger a standing WARNING about infrastructure
-    they do not own and cannot read.
+    Empty by default, and that default is why `main` refuses rather than
+    printing an empty board. Radar has no opinion about what anyone watches;
+    a tier that reaches for GitLab, or runners, or any other resource, must be
+    asked for by name.
 
     Nothing here raises. A config this cannot parse yields no tiers plus a
-    complaint, which renders as the ordinary board it was trying to extend.
+    complaint, which reaches the reader as the refusal plus the reason.
     """
     raw = os.environ.get(TIERS_ENV, "") if raw is None else raw
     raw = raw.strip()
@@ -579,7 +165,7 @@ def read_tiers(raw: str | None = None) -> tuple[dict[str, dict], list[str]]:
                     f"No tiers loaded."]
     if not isinstance(loaded, dict):
         return {}, ["radar: WARNING — radar_tiers must be an object keyed by op name, "
-                    "e.g. {'gl-runners': {}}. No tiers loaded."]
+                    "e.g. {'gl-mrs': {}}. No tiers loaded."]
 
     out: dict[str, dict] = {}
     problems: list[str] = []
@@ -590,17 +176,44 @@ def read_tiers(raw: str | None = None) -> tuple[dict[str, dict], list[str]]:
             problems.append(f"radar: WARNING — tier '{name}' options must be an object; "
                             f"got {type(opts).__name__}. Tier skipped.")
             continue
+        # Underscored keys are radar's own context channel (_arg, _watch). A
+        # config that could forge them would be a config that could lie to a
+        # tier about what radar asked for.
+        reserved = sorted(k for k in opts if str(k).startswith(RESERVED_PREFIX))
+        if reserved:
+            problems.append(f"radar: WARNING — tier '{name}' option(s) {reserved} start "
+                            f"with '{RESERVED_PREFIX}', which radar reserves for the "
+                            f"context it supplies. Dropped.")
+            opts = {k: v for k, v in opts.items()
+                    if not str(k).startswith(RESERVED_PREFIX)}
         out[str(name)] = opts
     return out, problems
 
 
 def _tier_module(name: str):
-    """Resolve a registered op name to the module implementing it, or None.
+    """Resolve a registered tier name to the module implementing it, or None.
 
-    Read out of the preset that declares the op rather than from a second
-    hardcoded table: the op's `cmd` already names its script, and a mapping
-    kept alongside it is one that drifts the first time a file moves.
+    Two places, in order:
+
+      1. `tiers/<name>.py` beside this file, with dashes as underscores. A tier
+         that needs radar's own internals — the transport, the dispatcher, the
+         shared watch defaults — belongs on this side of the preset boundary.
+         The MR board is that case: `presets/gitlab/mrs.py` is a GitLab preset
+         and the watch preset already depends on it, so putting reconcile
+         machinery there would make the dependency mutual.
+      2. the script the preset's op declares, read out of the preset itself
+         rather than from a second hardcoded table: the op's `cmd` already
+         names its script, and a mapping kept alongside it is one that drifts
+         the first time a file moves. `gl-runners` joins this way — its report
+         needs nothing but its own API helpers.
+
+    A name resolves to a directory entry or to an op, never to a table, so
+    neither route can fall out of step with the files on disk.
     """
+    local = _HERE / "tiers" / f"{name.replace('-', '_')}.py"
+    if local.is_file():
+        return _load(f"radar_tier_{name.replace('-', '_')}", local)
+
     presets_dir = _HERE.parent
     for preset in sorted(presets_dir.glob("*.json")):
         try:
@@ -619,23 +232,52 @@ def _tier_module(name: str):
     return None
 
 
-def tier_reports() -> tuple[list[str], bool]:
-    """(lines, all_healthy) from every registered tier, in registration order.
+def _spawner() -> tuple[Callable[..., str], dict[str, str]]:
+    """A `_watch` callable plus the ledger of every slot it was asked for.
 
-    A healthy tier is silent by default (`quiet_when_healthy`): the reader of a
-    board wants to know what is wrong, and a green line per tier per run is the
-    noise that trains people to skim past the red one. An unhealthy tier always
-    speaks, on every run, because it is a current fact rather than a transition.
+    The tier chooses when to spawn; radar keeps the bound and the record. A
+    slot radar refused is a slot nobody is polling, and the ledger is what lets
+    it say so on the same run.
+    """
+    seen: dict[str, str] = {}
+
+    def watch(source: str, scope: str, only: list[str] | None = None) -> str:
+        status = ensure_watcher(source, scope, only)
+        seen[f"{source}:{scope}"] = status
+        return status
+
+    return watch, seen
+
+
+def tier_reports(arg: str = "") -> tuple[list[str], bool, list[str]]:
+    """(lines, all_healthy, failures) from every registered tier, in order.
+
+    `lines` is the board — stdout. `failures` is the *cannot tell* channel: a
+    tier that could not be resolved, or that raised. It is returned apart from
+    `lines` rather than folded into them because the two have different
+    destinations and different exit codes. A broken tier must not be able to
+    cost a working one its board, and it must not be able to leave radar
+    exiting 0 either.
+
+    A healthy tier is silent when its `RADAR_QUIET_DEFAULT` says so — true for
+    a side concern, false for a tier whose report is the board itself.
+
+    Order is registration order, so an operator who wants the fleet verdict
+    above the MR board writes `gl-runners` first. Nothing here sorts: a tier
+    that says "the board below may be this, not your code" is making a claim
+    about position, and only the person who wrote the config can settle it.
     """
     tiers, lines = read_tiers()
+    watch, spawned = _spawner()
     all_ok = True
+    failures: list[str] = []
 
     for name, opts in tiers.items():
         module = _tier_module(name)
         report = getattr(module, "radar_report", None) if module else None
         if report is None:
-            lines.append(f"radar: WARNING — tier '{name}' is registered but exposes no "
-                         f"radar_report(); it contributes nothing. Check the name.")
+            failures.append(f"radar: WARNING — tier '{name}' is registered but exposes no "
+                            f"radar_report(); it contributes nothing. Check the name.")
             all_ok = False
             continue
 
@@ -645,405 +287,39 @@ def tier_reports() -> tuple[list[str], bool]:
                          f"{sorted(unknown)}; ignored. Check for a typo.")
 
         try:
-            tier_lines, ok = report(opts)
+            tier_lines, ok = report({**opts, "_arg": arg, "_watch": watch})
         except Exception as exc:  # noqa: BLE001 — a broken tier must not take radar down
-            lines.append(f"radar: WARNING — tier '{name}' failed: "
-                         f"{exc.__class__.__name__}: {exc}")
+            failures.append(f"radar: WARNING — tier '{name}' failed: "
+                            f"{exc.__class__.__name__}: {exc}")
             all_ok = False
             continue
 
         all_ok = all_ok and ok
-        if ok and opts.get("quiet_when_healthy", True):
+        quiet = opts.get("quiet_when_healthy",
+                         getattr(module, "RADAR_QUIET_DEFAULT", True))
+        if ok and quiet:
             continue
         lines.extend(tier_lines)
 
-    return lines, all_ok
-
-
-# ---------------------------------------------------------------------------
-# 4. report
-# ---------------------------------------------------------------------------
-
-def _snap_entry(m: dict) -> dict[str, Any]:
-    """The facts radar reports about one MR. Delta is computed over these."""
-    return {
-        "pipeline": str(m.get("_pipeline") or ""),
-        "pipeline_id": str(m.get("_pipeline_id") or ""),
-        "draft": bool(m.get("draft")),
-        # "conflict" | "empty" | "" — see mrs._conflict_label. This key held a
-        # bool before #471, so the first run after upgrading reads every row
-        # with a stored `false` as moved and prints a full board once.
-        "conflict": mrs._conflict_label(m),
-    }
-
-
-def read_snapshot(multi: dict[str, list[str]] | None = None) -> dict[str, Any] | None:
-    """Previous board for this filter, or None on cold start.
-
-    Keyed by filter: comparing one population against another's snapshot
-    reports every row of each as a change, which is a delta column that lies.
-    """
-    multi = default_filter() if multi is None else multi
-    try:
-        with open(_snapshot_path(multi), encoding="utf-8") as f:
-            loaded = json.load(f)
-    except (OSError, json.JSONDecodeError):
-        return None
-    if not isinstance(loaded, dict) or not isinstance(loaded.get("mrs"), dict):
-        return None
-    return loaded
-
-
-def write_snapshot(entries: dict[str, dict],
-                   multi: dict[str, list[str]] | None = None) -> None:
-    path = _snapshot_path(default_filter() if multi is None else multi)
-    tmp = f"{path}.tmp"
-    try:
-        with open(tmp, "w", encoding="utf-8") as f:
-            json.dump({"mrs": entries}, f, indent=2)
-        os.replace(tmp, path)
-    except OSError:
-        try:
-            os.unlink(tmp)
-        except OSError:
-            pass
-
-
-def _marks(iid: str, drifted: dict[str, tuple[str, str]],
-           healed: set[str], uncovered: set[str]) -> str:
-    """The two novel signals, appended to the shared gl-mrs row format."""
-    out = []
-    if iid in drifted:
-        was, now = drifted[iid]
-        out.append(f"[drift: {was}→{now}]")
-    if iid in healed:
-        out.append("[healed]")
-    elif iid in uncovered:
-        out.append("[unwatched]")
-    return ("  " + " ".join(out)) if out else ""
-
-
-def _is_standing_problem(m: dict) -> bool:
-    """Unresolved red or conflict — a current fact, so never delta-suppressed."""
-    return bool(_problem_label(m))
-
-
-def _problem_label(m: dict) -> str:
-    """"failed", "conflict", "empty", "failed+conflict", … or "" when fine.
-
-    Printed on the exclusion line, so a suppressed MR that picks up a second
-    problem changes what the board says about it without un-suppressing it.
-
-    An empty MR stays a standing problem, only under its own name: it is still
-    unmergeable, and dropping it out of the set would trade #471's mislabel
-    for a silent omission — the worse of the two (#445/#454/#414).
-    """
-    bits = []
-    if str(m.get("_pipeline") or "") == "failed":
-        bits.append("failed")
-    blocked = mrs._conflict_label(m)
-    if blocked:
-        bits.append(blocked)
-    return "+".join(bits)
-
-
-# ---------------------------------------------------------------------------
-# standing exclusions
-# ---------------------------------------------------------------------------
-
-def _refused(iid: str, why: str) -> str:
-    return (f"radar: exclusion !{iid} REFUSED — {why}. The row is shown; "
-            f"fix ops.radar.radar_exclusions in .supertool.json")
-
-
-def _not_applied(iid: str, why: str) -> str:
-    return f"radar: exclusion !{iid} NOT applied — {why}; {_FIX_HINT}"
-
-
-def read_exclusions(raw: str | None = None) -> tuple[dict[str, dict[str, str]], list[str]]:
-    """({iid: {"reason", "until"}}, complaints) from the configured JSON.
-
-    Shapes accepted, both keyed by iid: a bare reason string, or an object
-    with `reason` and optional `until` (ISO date). A reason is mandatory —
-    an exclusion nobody had to justify is the one that becomes a permanent
-    blind spot, and refusing it costs a line of output instead of a red.
-
-    Nothing here raises and nothing here defaults to suppression: a config
-    this function cannot understand yields no exclusions plus a complaint,
-    which renders as the ordinary board it was trying to trim.
-    """
-    raw = os.environ.get(EXCLUSIONS_ENV, "") if raw is None else raw
-    raw = raw.strip()
-    if not raw:
-        return {}, []
-    try:
-        loaded = json.loads(raw)
-    except json.JSONDecodeError as exc:
-        return {}, [f"radar: WARNING — radar_exclusions is not valid JSON ({exc.msg}). "
-                    f"No exclusions applied."]
-    if not isinstance(loaded, dict):
-        return {}, ["radar: WARNING — radar_exclusions must be an object keyed by MR "
-                    "iid. No exclusions applied."]
-
-    out: dict[str, dict[str, str]] = {}
-    problems: list[str] = []
-    for key, spec in loaded.items():
-        iid = str(key).strip().lstrip("!")
-        if isinstance(spec, str):
-            spec = {"reason": spec}
-        if not isinstance(spec, dict):
-            problems.append(_refused(iid, "it is neither a reason string nor an object"))
-            continue
-        if not iid.isdigit():
-            problems.append(_refused(iid, "the key is not an MR iid"))
-            continue
-        reason = str(spec.get("reason") or "").strip()
-        if not reason:
-            problems.append(_refused(iid, "it carries no reason"))
-            continue
-        out[iid] = {"reason": reason, "until": str(spec.get("until") or "").strip()}
-    return out, problems
-
-
-def resolve_exclusions(open_mrs: list[dict], exclusions: dict[str, dict[str, str]],
-                       covered: set[str], today: str = "") -> tuple[set[str], list[str]]:
-    """(suppressed iids, one accounting line per configured exclusion).
-
-    Matching is an exact iid lookup, never a prefix: excluding 1950 must not
-    swallow 19509.
-
-    An exclusion is honoured only while it is still true — the MR is in the
-    population, not past its `until`, and actually carrying a standing
-    problem. Each of the three failures prints the reason it did not apply
-    rather than silently doing nothing, because a standing exclusion that
-    stopped suppressing anything is dead config that will suppress something
-    the day that iid comes back.
-    """
-    if not exclusions:
-        return set(), []
-    today = today or datetime.date.today().isoformat()
-    by_iid = {str(m.get("iid")): m for m in open_mrs if m.get("iid") is not None}
-
-    suppressed: set[str] = set()
-    lines: list[str] = []
-    for iid in sorted(exclusions):
-        spec = exclusions[iid]
-        m = by_iid.get(iid)
-        if m is None:
-            lines.append(_not_applied(iid, "no open MR with that iid in this population"))
-            continue
-        until = spec["until"]
-        if until and until < today:
-            lines.append(_not_applied(iid, f"expired {until}"))
-            continue
-        label = _problem_label(m)
-        if not label:
-            pipe = str(m.get("_pipeline") or "none")
-            lines.append(_not_applied(
-                iid, f"pipeline is '{pipe}' and there is no conflict, so the "
-                     f"reason is spent"))
-            continue
-        suppressed.add(iid)
-        cover = "still watched" if iid in covered else "UNWATCHED"
-        lines.append(f"radar: excluded !{iid} {label}, {cover} — {spec['reason']}")
-    return suppressed, lines
-
-
-def _footer(open_mrs: list[dict], covered: set[str], healed: list[str],
-            drifted: dict[str, tuple[str, str]], pruned: list[str],
-            uncovered: list[str], gone: int, feed: str, label: str = "",
-            excluded: int = 0) -> str:
-    """Tallies over the board that was printed, plus what was kept off it.
-
-    `open_mrs` here is the *shown* population. A footer counting the full one
-    would report a failure with no row behind it, which sends the reader
-    hunting for something that was deliberately removed — the exclusion
-    restores the total as its own token instead.
-
-    `healed` / `unwatched` / `pruned` / `drift` stay over the whole
-    population: they report what radar did, and radar acts on excluded MRs
-    exactly as it acts on any other.
-    """
-    counts: dict[str, int] = {}
-    for m in open_mrs:
-        counts[str(m.get("_pipeline") or "none")] = counts.get(str(m.get("_pipeline") or "none"), 0) + 1
-    parts = [label] if label else []
-    parts.append(f"{len(open_mrs)} open")
-    if counts.get("failed"):
-        parts.append(f"{counts['failed']} failing")
-    if counts.get("running"):
-        parts.append(f"{counts['running']} running")
-    if counts.get("success"):
-        parts.append(f"{counts['success']} green")
-    parts.append(f"{len([m for m in open_mrs if str(m.get('iid')) in covered])} watched")
-    if healed:
-        parts.append(f"{len(healed)} healed")
-    if uncovered:
-        parts.append(f"{len(uncovered)} unwatched")
-    if drifted:
-        parts.append(f"{len(drifted)} drift")
-    if pruned:
-        parts.append(f"{len(pruned)} pruned")
-    if gone:
-        parts.append(f"{gone} no longer open")
-    if excluded:
-        parts.append(f"{excluded} excluded")
-    parts.append(FEED_LABEL.get(feed, feed))
-    return " | ".join(parts)
-
-
-def _feed_warnings(feed: str, feed_err: str,
-                   others: list[str] | None = None) -> list[str]:
-    """A blind radar must say so. A dead or erroring feed discovers nothing,
-    and the symptom of that is a board that simply stops gaining rows — which
-    is exactly what an all-quiet day looks like.
-
-    `others` is the same failure from the opposite side (#486): a feed left
-    running by an earlier, wider filter is still receiving MRs that this
-    board does not list, and the board looks healthy while omitting them.
-    Named, not killed — a second population can be deliberate.
-    """
-    out: list[str] = []
-    if feed == "failed":
-        out.append("radar: WARNING — MR feed poller is down. New MRs will not be "
-                   "discovered until the next radar run.")
-    elif feed_err:
-        out.append(f"radar: WARNING — MR feed poller is failing to poll: {feed_err}")
-    for other in others or []:
-        out.append(f"radar: NOTE — a feed poller is also live on scope '{other}', "
-                   f"which this board does not cover. Its MRs are not on this "
-                   f"board; re-run as radar:{other} to see them.")
-    return out
-
-
-def render(open_mrs: list[dict], covered: set[str], healed: list[str],
-           drifted: dict[str, tuple[str, str]], pruned: list[str],
-           uncovered: list[str], previous: dict[str, Any] | None,
-           feed: str = "alive", feed_err: str = "", label: str = "",
-           excluded: set[str] | None = None, notes: list[str] | None = None,
-           other_scopes: list[str] | None = None,
-           losses: list[str] | None = None,
-           fleet_lines: list[str] | None = None,
-           fleet_ok: bool = True) -> str:
-    """Full board on cold start; changed + standing-problem rows afterwards.
-
-    `label` names the population on every board, the default one included
-    (#486). Two radars over different filters in one window otherwise print
-    two boards indistinguishable from one board printed twice — and, worse,
-    a session that widened the filter and then ran a bare `radar` got the
-    narrow board back with nothing on it saying so. An omission the tool
-    produced renders exactly like an absence in the world, so the scope the
-    board was actually built from is stated rather than implied.
-
-    `excluded` removes rows; `notes` says so. They are two arguments rather
-    than one because the notes outlive the suppression — an exclusion that
-    did *not* apply prints a line and no row is removed, and that is the case
-    that stops a spent reason becoming a permanent blind spot. Notes are
-    printed on every run, including a delta-suppressed one: the repetition is
-    what the exclusion was removing, and making the suppression itself
-    invisible would rebuild the problem one level up.
-    """
-    cold = previous is None
-    prev_entries: dict[str, Any] = (previous or {}).get("mrs", {}) or {}
-    healed_set, uncovered_set = set(healed), set(uncovered)
-    excluded = excluded or set()
-    board_mrs = [m for m in open_mrs if str(m.get("iid", "?")) not in excluded]
-
-    shown = []
-    for m in sorted(board_mrs, key=mrs._sort_key):
-        iid = str(m.get("iid", "?"))
-        moved = prev_entries.get(iid) != _snap_entry(m)
-        notable = iid in drifted or iid in healed_set or iid in uncovered_set
-        if cold or moved or notable or _is_standing_problem(m):
-            marks = _marks(iid, drifted, healed_set, uncovered_set)
-            shown.append(mrs._row(m, covered, True, marks))
-
-    gone = len([i for i in prev_entries if i not in {str(m.get("iid")) for m in open_mrs}])
-    footer = _footer(board_mrs, covered, healed, drifted, pruned, uncovered, gone,
-                     feed, label, len(excluded))
-
-    lines = _feed_warnings(feed, feed_err, other_scopes) + list(losses or [])
-    # Printed as given. Tiers decide their own silence in `tier_reports` — a
-    # healthy tier already returned nothing unless it was asked to speak — so
-    # re-gating here swallowed two things that must never be swallowed: a tier
-    # explicitly configured to report when healthy, and the WARNING lines about
-    # a misconfigured tier, which are exactly the output a misconfiguration
-    # needs in order to get fixed.
-    if fleet_lines:
-        lines.extend(fleet_lines)
-    if cold:
-        lines.append("radar: cold start — no prior snapshot, full board")
-    if shown:
-        lines.extend(shown)
-        lines.append("")
-        lines.append(footer)
-    elif cold:
-        # "No open MRs" would be false when the population is non-empty and
-        # every row of it was excluded — the exact silent omission the
-        # accounting lines below exist to prevent.
-        lines.append("All open MRs in this population are excluded."
-                     if excluded else "No open MRs.")
-        lines.append("")
-        lines.append(footer)
-    else:
-        lines.append(f"radar: no change | {footer}")
-    lines.extend(notes or [])
-    return "\n".join(lines)
+    return watcher_cap_warnings(spawned) + lines, all_ok, failures
 
 
 def main(argv: list[str] | None = None) -> int:
-    multi = resolve_filter(argv)
-    try:
-        open_mrs = live_open_mrs(multi)
-    except RadarError as exc:
-        print(f"ERROR: {exc}", file=sys.stderr)
+    arg = argv[1].strip() if argv and len(argv) > 1 and argv[1] else ""
+
+    tiers, complaints = read_tiers()
+    if not tiers:
+        for line in complaints:
+            print(line, file=sys.stderr)
+        print(NO_TIERS, file=sys.stderr)
         return 1
 
-    open_iids = [str(m.get("iid")) for m in open_mrs if m.get("iid") is not None]
-    watched = mrs._watched_iids(transport.STATE_DIR)
-
-    states = read_state_files()
-    pruned = prune_terminal(states, watched)
-    drifted = drift({i: s for i, s in states.items() if i not in set(pruned)})
-
-    healed, uncovered, refused = heal(open_iids, watched)
-    covered = watched | set(healed)
-
-    scope = feed_scope(multi)
-    feed = ensure_feed(scope)
-    feed_err = feed_error(scope) if feed == "alive" else ""
-
-    registered, _tier_problems = read_tiers()
-    watcher_statuses: dict[str, str] = {}
-    for tier_name, (watch_source, watch_scope) in _TIER_WATCH_SOURCES.items():
-        if tier_name in registered:
-            watcher_statuses[f"{watch_source}:{watch_scope}"] = ensure_watcher(
-                watch_source, watch_scope)
-    fleet_lines, fleet_ok = tier_reports()
-    fleet_lines = watcher_cap_warnings(watcher_statuses) + fleet_lines
-
-    exclusions, excl_problems = read_exclusions()
-    excluded, excl_lines = resolve_exclusions(open_mrs, exclusions, covered)
-
-    # Stated on every board, default included: "no label" used to spell both
-    # "this is the default population" and "nobody said which population this
-    # is", and the filter does not survive an invocation (#486).
-    label = f"scope {filter_string(multi)}"
-    if multi == default_filter():
-        label += " (default)"
-    previous = read_snapshot(multi)
-    print(render(open_mrs, covered, healed, drifted, pruned, uncovered, previous,
-                 feed, feed_err, label, excluded, excl_problems + excl_lines,
-                 other_feed_scopes(scope), loss_warnings(healed, refused),
-                 fleet_lines, fleet_ok))
-    # The snapshot records the whole population, excluded rows included:
-    # keyed on what is true, not on what was printed. Otherwise the run after
-    # an exclusion is lifted reports a months-old MR as new.
-    write_snapshot(
-        {str(m.get("iid")): _snap_entry(m) for m in open_mrs if m.get("iid") is not None},
-        multi,
-    )
-    return 0
+    lines, _all_ok, failures = tier_reports(arg)
+    for line in failures:
+        print(line, file=sys.stderr)
+    if lines:
+        print("\n".join(lines))
+    return 1 if failures else 0
 
 
 if __name__ == "__main__":
